misc cleanup in host
